+++ pkg/core/constants.go
@@ -3,6 +3,6 @@ package core
 const ProgramName = "utdocs"
 const ProgramVersion = "1.0.0"
 const ThemeManifestName = "theme.ini"
-const SiteManifestName = ".env"
+const SiteManifestName = "n8go-docs.yaml"
 const RootTemplateName = "main.html"
 const IndexFileName = "index"